Refactor of the original code
BookTrader class was split :  abstract base class BookTraderBase implements only the communication protocol and the decision logic has to be written in a subclass. The same behaviour as the original (sample) implementation of the BookTrader is provided in the class RandomBookTrader.
New helper methods have been added into the BookTraderBase class, such as: getOwnedInstanceOfBook(book), getOwnedInstancesCount(book),  getMissingGoalBooks().